split file: move query specs into _query_spec.cpp file; minor refactoring
