+++ src/object/schema_system_catalog.hpp
@@ -31,6 +31,12 @@
 extern EXPORT_IMPORT void catcls_init (void);
 extern EXPORT_IMPORT int catcls_install (void);
 
+// test class_name is of system (class/vclass)s for legacy routine
+extern EXPORT_IMPORT bool sm_check_system_class_by_name (const std::string_view class_name);
+
+extern EXPORT_IMPORT bool sm_is_system_class (const std::string_view name);
+extern EXPORT_IMPORT bool sm_is_system_vclass (const std::string_view name);
+
 #if defined (CS_MODE) || defined (SA_MODE)
 struct db_object;
 
@@ -38,11 +44,5 @@ extern EXPORT_IMPORT int catcls_add_data_type (struct db_object *class_mop);
 extern EXPORT_IMPORT int catcls_add_charsets (struct db_object *class_mop);
 extern EXPORT_IMPORT int catcls_add_collations (struct db_object *class_mop);
 #endif
-
-// test class_name is of system (class/vclass)s for legacy routine
-extern EXPORT_IMPORT bool sm_check_system_class_by_name (const std::string_view class_name);
-
-extern EXPORT_IMPORT bool sm_is_system_class (const std::string_view name);
-extern EXPORT_IMPORT bool sm_is_system_vclass (const std::string_view name);
 
 #endif /* _SCHEMA_SYSTEM_CATALOG_HPP_ */
