+++ src/object/schema_system_catalog_builder.cpp
@@ -85,7 +85,9 @@ namespace cubschema
 	const char *name = attr.name.data ();
 	const char *type = attr.type.data ();
 
-	if (attr.kind == attribute_kind::COLUMN)
+	switch (attr.kind)
+	  {
+	  case attribute_kind::COLUMN:
 	  {
 	    error_code = smt_add_attribute (def, name, type, NULL);
 	    if (error_code != NO_ERROR)
@@ -101,13 +103,13 @@ namespace cubschema
 		error_code = smt_set_attribute_default (def, name, 0, (DB_VALUE *) &default_value, NULL);
 	      }
 	  }
-	else if (attr.kind == attribute_kind::CLASS_METHOD)
-	  {
+	  break;
+	  case attribute_kind::CLASS_METHOD:
 	    error_code = smt_add_class_method (def, name, type);
-	  }
-	else
-	  {
+	    break;
+	  default:
 	    error_code = ER_FAILED;
+	    break;
 	  }
 
 	if (error_code != NO_ERROR)
@@ -222,17 +224,17 @@ namespace cubschema
 	const char *name = attr.name.data ();
 	const char *type = attr.type.data ();
 
-	if (attr.kind == attribute_kind::COLUMN)
+	switch (attr.kind)
 	  {
+	  case attribute_kind::COLUMN:
 	    error_code = db_add_attribute (class_mop, name, type, NULL);
-	  }
-	else if (attr.kind == attribute_kind::QUERY_SPEC)
-	  {
-	    error_code = db_add_query_spec (class_mop, attr.name.data ());
-	  }
-	else
-	  {
+	    break;
+	  case attribute_kind::QUERY_SPEC:
+	    error_code = db_add_query_spec (class_mop, name);
+	    break;
+	  default:
 	    error_code = ER_FAILED;
+	    break;
 	  }
 
 	if (error_code != NO_ERROR)
